Create start and stop timer
This file consist of the code required to start or stop the timer as per user wish. 
The timer will input the current time and will stop when the button is clicked once.
The same button is then used to restart the time and the timer will display the current time again.